chore(wdio): port svg-class test suite to wdio
