--- test/karma/test-app/svg-class/index.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<meta charset="utf8">
-<script src="./build/testapp.esm.js" type="module"></script>
-<script src="./build/testapp.js" nomodule></script>
-
-<style>
-  .primary circle {
-    fill: red;
-  }
-  .primary rect {
-    fill: blue;
-  }
-</style>
-
-<svg-class></svg-class>
