+++ test/wdio/svg-class/cmp.tsx
@@ -1,4 +1,4 @@
-import { Component, State, h } from '@stencil/core';
+import { Component, h, State } from '@stencil/core';
 
 @Component({
   tag: 'svg-class',
